WIP: Add different trust levels to AppStore interface
:construction: :construction: :construction:

TODO:
- [ ] Have apps.owncloud.com return the trust levels – somehow
- [ ] Make it looks nicer
- [ ] Move the "Experimental applications ahead" bar right before the experimental apps
- [ ] Show the experimental applications at the end of the site if enabled
- [ ] Make the "Show experimental applications" button actually working

:construction: :construction: :construction:

Heads-up: We don't know when an application is activated (i.e. already installing) whether it is an official one / approved one or complete random stuff. Thus in the "activated" and "not activated" view we only show the official apps where we know it from the shipped flag and everything else has no flag applied.